fix split subjects

+++ brainlife-utils/data_utils.py
@@ -190,7 +190,7 @@ def split_subjects(sub_list, perc_val, seed=None):
     n_sub = len(sub_list)
     if seed != None:
         np.random.seed(seed)
-    n_sub_val = max(np.floor(perc_val * n_sub),1)
+    n_sub_val = max(int(np.floor(perc_val * n_sub)),1)
     sub_val = np.random.choice(sub_list, n_sub_val, replace=False).tolist()
     sub_train = [s for s in sub_list if s not in sub_val]
